api: renamed V2 API to V3 API for consistency.
